+++ src/components/index.ts
@@ -1,26 +1,14 @@
 export * from './common'
 export * from './button'
+export * from './board'
 
-export { BoardListMenu } from './board/menu/BoardListMenu'
-export { SearchView } from './board/menu/SearchView'
-export { BoardListView } from './board/menu/BoardListView'
-export { ToggleList } from './board/menu/ToggleList'
 export { BeforeVerified } from './views/BeforeVerified'
 export { BoardDetail } from './views/BoardDetail'
 export { BoardTop } from './views/BoardTop'
 export { Home } from './views/Home'
-export { ArchivedBoardModal } from './board/ArchivedBoardModal'
-export { BoardListSearchForm } from './board/BoardListSearchForm'
-export { BoardListItem } from './board/BoardListItem'
-export { BoardTitle } from './board/BoardTitle'
-export { BoardWithBackground } from './board/BoardWithBackground'
-export { CreateBoardModal } from './board/CreateBoardModal'
 export { CardList } from './list/CardList'
 export { ListContainer } from './list/ListContainer'
-export { DrawerSelectBg } from './board/drawer/DrawerSelectBg'
-export { BoardDrawer } from './board/drawer/BoardDrawer'
-export { DrawerArchivedItem } from './board/drawer/DrawerArchivedItem'
-export { DrawerRoot } from './board/drawer/DrawerRoot'
+
 export { EMailField } from './auth/EMailField'
 export { LoginView } from './auth/LoginView'
 export { LoginFormModal } from './auth/LoginFormModal'
