+++ src/scripts/context/CreateBoardModalContext.ts
@@ -1,0 +1,9 @@
+import { createContext } from 'react'
+
+export type ContextType = {
+  openCreateBoardModal: () => void
+}
+
+export const Context = createContext<ContextType>({
+  openCreateBoardModal: () => undefined
+})
