--- src/scripts/context/index.ts
@@ -1,1 +1,0 @@
-export * from './modules/SnackbarContext'
