+++ src/scripts/hooks/index.ts
@@ -2,6 +2,7 @@ export { useAuth } from './modules/useAuth'
 export { useEventListener } from './modules/useEventListener'
 export { useMountedRef } from './modules/useMountedRef'
 export { useSnackbarContext } from './modules/useSnackbarContext'
+export { useCreateBoardModalContext } from './modules/useCreateBoardModalContext'
 export { usePrevious } from './modules/usePrevious'
 export { useFetchBoard } from './modules/useFetchBoard'
 export { useFetchBoards } from './modules/useFetchBoards'
