+++ src/scripts/hooks/modules/useSnackbarContext.ts
@@ -1,11 +1,9 @@
 import { useContext } from 'react'
-import * as Snackbar from '@/scripts/context'
+import * as Snackbar from '@/scripts/context/SnackbarContext'
 
 /**
  * ポップアップ表示に使うメソッドを提供するカスタムフック
  * SnackbarContext を利用するシンタックスシュガー
  */
-export const useSnackbarContext = (): Snackbar.ContextType => {
-  const { showSnackbar, closeSnackbar } = useContext(Snackbar.Context)
-  return { showSnackbar, closeSnackbar }
-}
+export const useSnackbarContext = (): Snackbar.ContextType =>
+  useContext(Snackbar.Context)
